Add raw data
This data is to be kept in a separate branch - it's only
useful as an intermediate step in creating the parsed data,
and it is much bigger than the parsed data.

It can also be reaquired from the NRA website, though that
takes a few hours as of writing this.